+++ mobile/android/android-components/components/browser/toolbar/src/test/java/mozilla/components/browser/toolbar/BrowserToolbarTest.kt
@@ -743,6 +743,45 @@ class BrowserToolbarTest {
     }
 
     @Test
+    fun `titleTextSize changes display titleView`() {
+        val toolbar = BrowserToolbar(context)
+
+        assertTrue(toolbar.displayToolbar.titleView.textSize != 12f)
+
+        toolbar.titleTextSize = 12f
+
+        assertEquals(12f, toolbar.displayToolbar.titleView.textSize)
+    }
+
+    @Test
+    fun `titleTextColor changes display titleView`() {
+        val toolbar = BrowserToolbar(context)
+
+        toolbar.titleColor = R.color.photonBlue40
+
+        assertEquals(R.color.photonBlue40, toolbar.displayToolbar.titleView.currentTextColor)
+    }
+
+    @Test
+    fun `titleView visibility is based on being set`() {
+        val toolbar = BrowserToolbar(context)
+
+        assertEquals(toolbar.displayToolbar.titleView.visibility, View.GONE)
+        toolbar.title = "Mozilla"
+        assertEquals(toolbar.displayToolbar.titleView.visibility, View.VISIBLE)
+        toolbar.title = ""
+        assertEquals(toolbar.displayToolbar.titleView.visibility, View.GONE)
+    }
+
+    @Test
+    fun `titleView text is set properly`() {
+        val toolbar = BrowserToolbar(context)
+
+        toolbar.title = "Mozilla"
+        assertEquals(toolbar.displayToolbar.titleView.text, "Mozilla")
+    }
+
+    @Test
     fun `typeface changes edit and display urlView`() {
         val toolbar = BrowserToolbar(context)
         val typeface: Typeface = mock()