+++ mobile/android/android-components/components/feature/customtabs/src/main/java/mozilla/components/feature/customtabs/CustomTabsToolbarFeature.kt
@@ -43,7 +43,10 @@ class CustomTabsToolbarFeature(
         if (initialized) {
             return
         }
-        initialized = sessionManager.runWithSession(sessionId) { initialize(it) }
+        initialized = sessionManager.runWithSession(sessionId) {
+            it.register(sessionObserver)
+            initialize(it)
+        }
     }
 
     @VisibleForTesting
@@ -65,6 +68,10 @@ class CustomTabsToolbarFeature(
             if (config.showShareMenuItem) addShareButton(session)
             // Add menu items
             if (config.menuItems.isNotEmpty()) addMenuItems(config.menuItems)
+
+            // Explicitly set the title regardless of the customTabConfig settings
+            toolbar.titleTextSize = TITLE_TEXT_SIZE
+
             return true
         }
         return false
@@ -75,6 +82,7 @@ class CustomTabsToolbarFeature(
         toolbarColor?.let { color ->
             toolbar.setBackgroundColor(color)
             toolbar.textColor = readableColor
+            toolbar.titleColor = readableColor
             toolbar.siteSecurityColor = Pair(readableColor, readableColor)
             toolbar.menuViewColor = readableColor
         }
@@ -134,7 +142,20 @@ class CustomTabsToolbarFeature(
         }
     }
 
-    override fun stop() {}
+    private val sessionObserver = object : Session.Observer {
+        override fun onTitleChanged(session: Session, title: String) {
+            // Only shrink the urlTextSize if a title is displayed
+            toolbar.textSize = URL_TEXT_SIZE
+            toolbar.title = title
+        }
+    }
+
+    override fun stop() {
+        sessionManager.runWithSession(sessionId) {
+            it.unregister(sessionObserver)
+            true
+        }
+    }
 
     /**
      * When the back button is pressed if not initialized returns false,
@@ -154,6 +175,8 @@ class CustomTabsToolbarFeature(
     }
 
     companion object {
+        const val TITLE_TEXT_SIZE = 16f
+        const val URL_TEXT_SIZE = 12f
         @Suppress("MagicNumber")
         internal fun getReadableTextColor(backgroundColor: Int): Int {
             val greyValue = greyscaleFromRGB(backgroundColor)